Refactor chunk sync codes.
Related codes are moved to a separate module.
