+++ core/src/sync/mod.rs
@@ -18,7 +18,7 @@ pub mod utils;
 
 pub use self::{
     error::{Error, ErrorKind},
-    state::{restore, StateSyncConfiguration},
+    state::StateSyncConfiguration,
     synchronization_graph::{
         SharedSynchronizationGraph, SyncGraphConfig, SyncGraphStatistics,
         SynchronizationGraph, SynchronizationGraphInner,
