+++ core/src/sync/state/mod.rs
@@ -2,23 +2,12 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-pub mod restore;
-mod snapshot_chunk_request;
-mod snapshot_chunk_response;
 mod snapshot_chunk_sync;
-mod snapshot_manifest_request;
-mod snapshot_manifest_response;
-mod state_sync_candidate_manager;
-mod state_sync_candidate_request;
-mod state_sync_candidate_response;
-mod storage;
+mod state_sync_candidate;
+mod state_sync_chunk;
+mod state_sync_manifest;
+pub mod storage;
 
-pub use self::{
-    snapshot_chunk_request::SnapshotChunkRequest,
-    snapshot_chunk_response::SnapshotChunkResponse,
-    snapshot_chunk_sync::{SnapshotChunkSync, StateSyncConfiguration, Status},
-    snapshot_manifest_request::SnapshotManifestRequest,
-    snapshot_manifest_response::SnapshotManifestResponse,
-    state_sync_candidate_request::StateSyncCandidateRequest,
-    state_sync_candidate_response::StateSyncCandidateResponse,
+pub use self::snapshot_chunk_sync::{
+    SnapshotChunkSync, StateSyncConfiguration, Status,
 };
