+++ core/src/sync/state/state_sync_manifest/mod.rs
@@ -1,0 +1,1 @@
+pub mod snapshot_manifest_manager;
